Refactor context IDs
