+++ Sources/WebAPIKit/Support.swift
@@ -23,49 +23,3 @@ public typealias HTMLOrSVGScriptElement = HTMLScriptElement
 public typealias BodyInit = XMLHttpRequestBodyInit
 public typealias CustomElementConstructor = JSFunction
 public typealias GLintptr = Int32
-
-public protocol RenderingContext: JSValueCompatible {}
-
-extension GPUCanvasContext: RenderingContext {}
-extension ImageBitmapRenderingContext: RenderingContext {}
-extension OffscreenCanvasRenderingContext2D: RenderingContext {}
-extension WebGL2RenderingContext: RenderingContext {}
-extension WebGLRenderingContext: RenderingContext {}
-
-public protocol OffscreenRenderingContext: JSValueCompatible {}
-
-extension GPUCanvasContext: OffscreenRenderingContext {}
-extension ImageBitmapRenderingContext: OffscreenRenderingContext {}
-extension OffscreenCanvasRenderingContext2D: OffscreenRenderingContext {}
-extension WebGL2RenderingContext: OffscreenRenderingContext {}
-extension WebGLRenderingContext: OffscreenRenderingContext {}
-
-extension Strings {
-    @usableFromInline static let getContext: JSString = "getContext"
-}
-
-public extension HTMLCanvasElement {
-    @inlinable func getContext<Context: RenderingContext>(
-        contextId: String,
-        options: JSValue? = nil
-    ) -> Context? {
-        let this = jsObject
-        return this[Strings.getContext].function!(
-            this: this,
-            arguments: [contextId.jsValue, options?.jsValue ?? .undefined]
-        ).fromJSValue()!
-    }
-}
-
-public extension OffscreenCanvas {
-    @inlinable func getContext<Context: OffscreenRenderingContext>(
-        contextId: OffscreenRenderingContextId,
-        options: JSValue? = nil
-    ) -> Context? {
-        let this = jsObject
-        return this[Strings.getContext].function!(
-            this: this,
-            arguments: [contextId.jsValue, options?.jsValue ?? .undefined]
-        ).fromJSValue()!
-    }
-}
