+++ Sources/WebAPIKitDemo/WebGLDemo.swift
@@ -117,7 +117,7 @@ func runWebGLDemo() {
   // Get A WebGL context
   let canvas = HTMLCanvasElement(from: document.createElement(localName: "canvas"))!
   _ = document.body?.appendChild(node: canvas)
-  guard let context: WebGL2RenderingContext = canvas.getContext(contextId: "webgl2") else {
+  guard let context = canvas.getContext(WebGL2RenderingContext.self) else {
     console.error(data: "Failed to create WebGL2 rendering context")
     return
   }
